+++ docs/changes.rst
@@ -32,7 +32,9 @@ v1.1.3 (in development)
 
 **Crawlers**
 
+- New: ``extraordinary``
 - Update ``eatthattoast`` after site change.
+- Update ``questionablecontent`` to be more robust.
 - Update ``threewordphrase`` to be more robust.